+++ Dockerfile
@@ -18,15 +18,15 @@ RUN apt install -y meshlab
 RUN cd /root ; mkdir libIGEVStereo
 RUN cd /root/libIGEVStereo
 WORKDIR /root/libIGEVStereo
-RUN mkdir -p /root/libIGEVStereo/libigev_stereo/
+RUN mkdir -p /root/libIGEVStereo/stereoigev/
 RUN mkdir /root/libIGEVStereo/scripts/
 RUN mkdir -p /root/libIGEVStereo/test/test-imgs/
-COPY libigev_stereo/*.py /root/libIGEVStereo/libigev_stereo/
+COPY stereoigev/*.py /root/libIGEVStereo/stereoigev/
 COPY *.py ./
 COPY test/test-imgs/ /root/libIGEVStereo/test/test-imgs/
 COPY test/*.py test/*.sh /root/libIGEVStereo/test/
 RUN python3 -m pip install gdown
-RUN mkdir -p /root/libIGEVStereo/libigev_stereo/models/ ; cd /root/libIGEVStereo/libigev_stereo/models/ ; gdown --fuzzy https://drive.google.com/file/d/16e9NR_RfzFdYT5mPaGwpjccplCi82C2e/view?usp=drive_link
+RUN mkdir -p /root/libIGEVStereo/stereoigev/models/ ; cd /root/libIGEVStereo/stereoigev/models/ ; gdown --fuzzy https://drive.google.com/file/d/16e9NR_RfzFdYT5mPaGwpjccplCi82C2e/view?usp=drive_link
 COPY pyproject.toml Makefile ./
 COPY sample.sh ./
