+++ sample.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
-python3 demo_imgs.py --restore_ckpt ./libigev_stereo/models/sceneflow.pth -l test/test-imgs/PlaytableP/im0.png -r test/test-imgs/PlaytableP/im1.png
+python3 demo_imgs.py --restore_ckpt ./stereoigev/models/sceneflow.pth -l test/test-imgs/PlaytableP/im0.png -r test/test-imgs/PlaytableP/im1.png
