+++ stereoigev/igev_stereo.py
@@ -2,10 +2,10 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from libigev_stereo.update import BasicMultiUpdateBlock
-from libigev_stereo.extractor import MultiBasicEncoder, Feature
-from libigev_stereo.geometry import Combined_Geo_Encoding_Volume
-from libigev_stereo.submodule import (
+from stereoigev.update import BasicMultiUpdateBlock
+from stereoigev.extractor import MultiBasicEncoder, Feature
+from stereoigev.geometry import Combined_Geo_Encoding_Volume
+from stereoigev.submodule import (
     BasicConv,
     BasicConv_IN,
     Conv2x_IN,
