+++ stereoigev/utils.py
